@@ -1,11 +1,10 @@
 name: table_calendar
 description: Highly customizable, feature-packed Flutter Calendar with gestures, animations and multiple formats.
 version: 2.2.3
-author: Aleksander Woźniak <aleksanderwozniak96@gmail.com>
 homepage: https://github.com/aleksanderwozniak/table_calendar
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -13,6 +12,7 @@ dependencies:
   
   intl: ^0.17.0
   simple_gesture_detector: ^0.2.0
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   flutter_test:
