@@ -13,10 +13,10 @@ typedef void OnVisibleDaysChanged(DateTime first, DateTime last, CalendarFormat 
 typedef void OnCalendarCreated(DateTime first, DateTime last, CalendarFormat format);
 
 /// Signature for reacting to header gestures. Exposes current month and year as a `DateTime` object.
-typedef void HeaderGestureCallback(DateTime focusedDay);
+typedef void HeaderGestureCallback(DateTime? focusedDay);
 
 /// Builder signature for any text that can be localized and formatted with `DateFormat`.
-typedef String TextBuilder(DateTime date, dynamic locale);
+typedef String TextBuilder(DateTime? date, dynamic locale);
 
 /// Signature for enabling days.
 typedef bool EnabledDayPredicate(DateTime day);
@@ -57,48 +57,48 @@ class TableCalendar extends StatefulWidget {
 
   /// `Map` of events.
   /// Each `DateTime` inside this `Map` should get its own `List` of objects (i.e. events).
-  final Map<DateTime, List> events;
+  final Map<DateTime, List>? events;
 
   /// `Map` of holidays.
   /// This property allows you to provide custom holiday rules.
   final Map<DateTime, List> holidays;
 
   /// Called whenever any day gets tapped.
-  final OnDaySelected onDaySelected;
+  final OnDaySelected? onDaySelected;
 
   /// Called whenever any day gets long pressed.
-  final OnDaySelected onDayLongPressed;
+  final OnDaySelected? onDayLongPressed;
 
   /// Called whenever any unavailable day gets tapped.
   /// Replaces `onDaySelected` for those days.
-  final VoidCallback onUnavailableDaySelected;
+  final VoidCallback? onUnavailableDaySelected;
 
   /// Called whenever any unavailable day gets long pressed.
   /// Replaces `onDaySelected` for those days.
-  final VoidCallback onUnavailableDayLongPressed;
+  final VoidCallback? onUnavailableDayLongPressed;
 
   /// Called whenever header gets tapped.
-  final HeaderGestureCallback onHeaderTapped;
+  final HeaderGestureCallback? onHeaderTapped;
 
   /// Called whenever header gets long pressed.
-  final HeaderGestureCallback onHeaderLongPressed;
+  final HeaderGestureCallback? onHeaderLongPressed;
 
   /// Called whenever the range of visible days changes.
-  final OnVisibleDaysChanged onVisibleDaysChanged;
+  final OnVisibleDaysChanged? onVisibleDaysChanged;
 
   /// Called once when the CalendarController gets initialized.
-  final OnCalendarCreated onCalendarCreated;
+  final OnCalendarCreated? onCalendarCreated;
 
   /// Initially selected DateTime. Usually it will be `DateTime.now()`.
-  final DateTime initialSelectedDay;
+  final DateTime? initialSelectedDay;
 
   /// The first day of `TableCalendar`.
   /// Days before it will use `unavailableStyle` and run `onUnavailableDaySelected` callback.
-  final DateTime startDay;
+  final DateTime? startDay;
 
   /// The last day of `TableCalendar`.
   /// Days after it will use `unavailableStyle` and run `onUnavailableDaySelected` callback.
-  final DateTime endDay;
+  final DateTime? endDay;
 
   /// List of days treated as weekend days.
   /// Use built-in `DateTime` weekday constants (e.g. `DateTime.monday`) instead of `int` literals (e.q. `1`).
@@ -126,10 +126,10 @@ class TableCalendar extends StatefulWidget {
 
   /// Function deciding whether given day should be enabled or not.
   /// If `false` is returned, this day will be unavailable.
-  final EnabledDayPredicate enabledDayPredicate;
+  final EnabledDayPredicate? enabledDayPredicate;
 
   /// Used for setting the height of `TableCalendar`'s rows.
-  final double rowHeight;
+  final double? rowHeight;
 
   /// Animation to run when `CalendarFormat` gets changed.
   final FormatAnimation formatAnimation;
@@ -162,8 +162,8 @@ class TableCalendar extends StatefulWidget {
   final CalendarBuilders builders;
 
   TableCalendar({
-    Key key,
-    @required this.calendarController,
+    Key? key,
+    required this.calendarController,
     this.locale,
     this.events = const {},
     this.holidays = const {},
@@ -200,10 +200,8 @@ class TableCalendar extends StatefulWidget {
     this.daysOfWeekStyle = const DaysOfWeekStyle(),
     this.headerStyle = const HeaderStyle(),
     this.builders = const CalendarBuilders(),
-  })  : assert(calendarController != null),
-        assert(availableCalendarFormats.keys.contains(initialCalendarFormat)),
+  })  : assert(availableCalendarFormats.keys.contains(initialCalendarFormat)),
         assert(availableCalendarFormats.length <= CalendarFormat.values.length),
-        assert(weekendDays != null),
         assert(weekendDays.isNotEmpty
             ? weekendDays.every((day) => day >= DateTime.monday && day <= DateTime.sunday)
             : true),
@@ -248,7 +246,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
 
   void _selectedDayCallback(DateTime day) {
     if (widget.onDaySelected != null) {
-      widget.onDaySelected(day, widget.calendarController.visibleEvents[_getEventKey(day)] ?? []);
+      widget.onDaySelected!(day, widget.calendarController.visibleEvents[_getEventKey(day)!] ?? []);
     }
   }
 
@@ -273,7 +271,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
 
   void _onDayLongPressed(DateTime day) {
     if (widget.onDayLongPressed != null) {
-      widget.onDayLongPressed(day, widget.calendarController.visibleEvents[_getEventKey(day)] ?? []);
+      widget.onDayLongPressed!(day, widget.calendarController.visibleEvents[_getEventKey(day)!] ?? []);
     }
   }
 
@@ -295,43 +293,43 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
 
   void _onUnavailableDaySelected() {
     if (widget.onUnavailableDaySelected != null) {
-      widget.onUnavailableDaySelected();
+      widget.onUnavailableDaySelected!();
     }
   }
 
   void _onUnavailableDayLongPressed() {
     if (widget.onUnavailableDayLongPressed != null) {
-      widget.onUnavailableDayLongPressed();
+      widget.onUnavailableDayLongPressed!();
     }
   }
 
   void _onHeaderTapped() {
     if (widget.onHeaderTapped != null) {
-      widget.onHeaderTapped(widget.calendarController.focusedDay);
+      widget.onHeaderTapped!(widget.calendarController.focusedDay);
     }
   }
 
   void _onHeaderLongPressed() {
     if (widget.onHeaderLongPressed != null) {
-      widget.onHeaderLongPressed(widget.calendarController.focusedDay);
+      widget.onHeaderLongPressed!(widget.calendarController.focusedDay);
     }
   }
 
   bool _isDayUnavailable(DateTime day) {
-    return (widget.startDay != null && day.isBefore(widget.calendarController._normalizeDate(widget.startDay))) ||
-        (widget.endDay != null && day.isAfter(widget.calendarController._normalizeDate(widget.endDay))) ||
+    return (widget.startDay != null && day.isBefore(widget.calendarController._normalizeDate(widget.startDay!))) ||
+        (widget.endDay != null && day.isAfter(widget.calendarController._normalizeDate(widget.endDay!))) ||
         (!_isDayEnabled(day));
   }
 
   bool _isDayEnabled(DateTime day) {
-    return widget.enabledDayPredicate == null ? true : widget.enabledDayPredicate(day);
+    return widget.enabledDayPredicate == null ? true : widget.enabledDayPredicate!(day);
   }
 
-  DateTime _getEventKey(DateTime day) {
+  DateTime? _getEventKey(DateTime day) {
     return widget.calendarController._getEventKey(day);
   }
 
-  DateTime _getHolidayKey(DateTime day) {
+  DateTime? _getHolidayKey(DateTime day) {
     return widget.calendarController._getHolidayKey(day);
   }
 
@@ -365,8 +363,8 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
           onLongPress: _onHeaderLongPressed,
           child: Text(
             widget.headerStyle.titleTextBuilder != null
-                ? widget.headerStyle.titleTextBuilder(widget.calendarController.focusedDay, widget.locale)
-                : DateFormat.yMMMM(widget.locale).format(widget.calendarController.focusedDay),
+                ? widget.headerStyle.titleTextBuilder!(widget.calendarController.focusedDay, widget.locale)
+                : DateFormat.yMMMM(widget.locale).format(widget.calendarController.focusedDay!),
             style: widget.headerStyle.titleTextStyle,
             textAlign: widget.headerStyle.centerHeaderTitle ? TextAlign.center : TextAlign.start,
           ),
@@ -403,7 +401,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
         decoration: widget.headerStyle.formatButtonDecoration,
         padding: widget.headerStyle.formatButtonPadding,
         child: Text(
-          widget.calendarController._getFormatButtonText(),
+          widget.calendarController._getFormatButtonText()!,
           style: widget.headerStyle.formatButtonTextStyle,
         ),
       ),
@@ -438,7 +436,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     }
   }
 
-  Widget _buildWrapper({Key key}) {
+  Widget _buildWrapper({Key? key}) {
     Widget wrappedChild = _buildTable();
 
     switch (widget.availableGestures) {
@@ -469,7 +467,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     );
   }
 
-  Widget _buildVerticalSwipeWrapper({Widget child}) {
+  Widget _buildVerticalSwipeWrapper({required Widget child}) {
     return SimpleGestureDetector(
       child: child,
       onVerticalSwipe: (direction) {
@@ -481,7 +479,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     );
   }
 
-  Widget _buildHorizontalSwipeWrapper({Widget child}) {
+  Widget _buildHorizontalSwipeWrapper({required Widget child}) {
     return AnimatedSwitcher(
       duration: const Duration(milliseconds: 350),
       switchInCurve: Curves.decelerate,
@@ -492,7 +490,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
           child: child,
         );
       },
-      layoutBuilder: (currentChild, _) => currentChild,
+      layoutBuilder: (currentChild, _) => currentChild!,
       child: Dismissible(
         key: ValueKey(widget.calendarController._pageId),
         resizeDuration: null,
@@ -510,8 +508,8 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     ];
 
     int x = 0;
-    while (x < widget.calendarController._visibleDays.value.length) {
-      children.add(_buildTableRow(widget.calendarController._visibleDays.value.skip(x).take(daysInWeek).toList()));
+    while (x < widget.calendarController._visibleDays!.value.length) {
+      children.add(_buildTableRow(widget.calendarController._visibleDays!.value.skip(x).take(daysInWeek).toList()));
       x += daysInWeek;
     }
 
@@ -524,17 +522,17 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
 
   TableRow _buildDaysOfWeek() {
     return TableRow(
-      children: widget.calendarController._visibleDays.value.take(7).map((date) {
+      children: widget.calendarController._visibleDays!.value.take(7).map((date) {
         final weekdayString = widget.daysOfWeekStyle.dowTextBuilder != null
-            ? widget.daysOfWeekStyle.dowTextBuilder(date, widget.locale)
+            ? widget.daysOfWeekStyle.dowTextBuilder!(date, widget.locale)
             : DateFormat.E(widget.locale).format(date);
         final isWeekend = widget.calendarController._isWeekend(date, widget.weekendDays);
 
         if (isWeekend && widget.builders.dowWeekendBuilder != null) {
-          return widget.builders.dowWeekendBuilder(context, weekdayString);
+          return widget.builders.dowWeekendBuilder!(context, weekdayString);
         }
         if (widget.builders.dowWeekdayBuilder != null) {
-          return widget.builders.dowWeekdayBuilder(context, weekdayString);
+          return widget.builders.dowWeekdayBuilder!(context, weekdayString);
         }
         return Center(
           child: Text(
@@ -584,7 +582,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
       if (!_isDayUnavailable(date)) {
         if (widget.builders.markersBuilder != null) {
           children.addAll(
-            widget.builders.markersBuilder(
+            widget.builders.markersBuilder!(
               context,
               key,
               events,
@@ -600,7 +598,7 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
               right: widget.calendarStyle.markersPositionRight,
               child: Row(
                 mainAxisSize: MainAxisSize.min,
-                children: events
+                children: events!
                     .take(widget.calendarStyle.markersMaxAmount)
                     .map((event) => _buildMarker(eventKey, event))
                     .toList(),
@@ -648,25 +646,25 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     final isWeekend = widget.builders.weekendDayBuilder != null && !tIsOutside && tIsWeekend && !tIsHoliday;
 
     if (isUnavailable) {
-      return widget.builders.unavailableDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.unavailableDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isSelected && widget.calendarStyle.renderSelectedFirst) {
-      return widget.builders.selectedDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.selectedDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isToday) {
-      return widget.builders.todayDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.todayDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isSelected) {
-      return widget.builders.selectedDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.selectedDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isOutsideHoliday) {
-      return widget.builders.outsideHolidayDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.outsideHolidayDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isHoliday) {
-      return widget.builders.holidayDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.holidayDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isOutsideWeekend) {
-      return widget.builders.outsideWeekendDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.outsideWeekendDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isOutside) {
-      return widget.builders.outsideDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.outsideDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (isWeekend) {
-      return widget.builders.weekendDayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.weekendDayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else if (widget.builders.dayBuilder != null) {
-      return widget.builders.dayBuilder(context, date, widget.calendarController.visibleEvents[eventKey]);
+      return widget.builders.dayBuilder!(context, date, widget.calendarController.visibleEvents[eventKey!]);
     } else {
       return _CellWidget(
         text: '${date.day}',
@@ -681,9 +679,9 @@ class _TableCalendarState extends State<TableCalendar> with SingleTickerProvider
     }
   }
 
-  Widget _buildMarker(DateTime date, dynamic event) {
+  Widget _buildMarker(DateTime? date, dynamic event) {
     if (widget.builders.singleMarkerBuilder != null) {
-      return widget.builders.singleMarkerBuilder(context, date, event);
+      return widget.builders.singleMarkerBuilder!(context, date, event);
     } else {
       return Container(
         width: 8.0,
