@@ -14,18 +14,16 @@ class _CellWidget extends StatelessWidget {
   final CalendarStyle calendarStyle;
 
   const _CellWidget({
-    Key key,
-    @required this.text,
+    Key? key,
+    required this.text,
     this.isUnavailable = false,
     this.isSelected = false,
     this.isToday = false,
     this.isWeekend = false,
     this.isOutsideMonth = false,
     this.isHoliday = false,
-    @required this.calendarStyle,
-  })  : assert(text != null),
-        assert(calendarStyle != null),
-        super(key: key);
+    required this.calendarStyle,
+  })  : super(key: key);
 
   @override
   Widget build(BuildContext context) {
